Added alert when creating empty item and input cleaner function

+++ starter-code/js/index.js
@@ -39,14 +39,27 @@ window.onload = function(){
 
   function createNewItem(){
     var totalItemList = document.getElementById("totalitemlist");
-    var newItemDiv = initialItem.cloneNode(true);
     var newItem = document.getElementById("newitem").value;
     var newPrice = document.getElementById("newprice").value;
-    newItemDiv.childNodes[1].innerHTML = newItem;
-    newItemDiv.childNodes[3].childNodes[3].innerHTML = newPrice;
-    newItemDiv.childNodes[9].onclick = function() {
-      deleteItem(this);
-    };
-    totalItemList.appendChild(newItemDiv);
+    // Check if the input fields are empty.
+    if (newItem && newPrice) {
+      var newItemDiv = initialItem.cloneNode(true);
+      // Change every childNodes from the cloned element.
+      newItemDiv.childNodes[1].innerHTML = newItem;
+      newItemDiv.childNodes[3].childNodes[3].innerHTML = newPrice;
+      newItemDiv.childNodes[7].childNodes[3].innerHTML = 0.00;
+      newItemDiv.childNodes[5].childNodes[3].value = "";
+      newItemDiv.childNodes[9].onclick = function() {
+        deleteItem(this);
+      };
+      // Append the child elements created.
+      totalItemList.appendChild(newItemDiv);
+      // Eliminate the actual value of the inputs used.
+      document.getElementById("newitem").value = "";
+      document.getElementById("newprice").value = "";
+    } else {
+      alert("Input fields are empty! Can't create a new item!");
+    }
+
   }
 };
